Remove LZO typedtext stuff from scalding-core
