+++ scalding-commons/src/test/scala/com/twitter/scalding/commons/source/typedtext/TypedTextTest.scala
@@ -1,6 +1,7 @@
-package com.twitter.scalding.source.typedtext
+package com.twitter.scalding.commons.source.lzotypedtext
 
 import com.twitter.scalding._
+import com.twitter.scalding.source.typedtext._
 
 import org.scalatest.FunSuite
 import org.scalatest.junit.JUnitRunner
@@ -9,13 +10,8 @@ case class Test1(a: Int, b: Long, c: Option[Double])
 case class Test2(one: Test1, d: String)
 
 class TypedTextTest extends FunSuite {
-  test("Test with a flat tuple") {
-    val source = TypedText.tsv[Test1]("myPath")
-    assert(source.sourceFields.size == 3)
-  }
-
-  test("Test with a nested tuple") {
-    val source = TypedText.tsv[Test2]("myPath")
+  test("Test with a nested tuple: Daily") {
+    val source = LzoTypedText.dailyLzoTsv[Test2]("myPath")(DateRange(RichDate.now, RichDate.now + Hours(1)), implicitly)
     assert(source.sourceFields.size == 4)
   }
 }
